feat: introduce assets generation
Introduce ``generate-icons`` and ``generate-splashscreens`` commands to generate assets needed for iOS and Android.
The operations are as follow:
 - Icons:
  - For all platforms we just resize the provided image.
 - Splash screens:
  - For Android we generate 2 png for each dpi: blank with background(background.png) and resized image based on provided image(logo.png). During the app start the logo overlays the background.
  - For iOS we generate 1 png for each dpi which has resized and centered image on the background.

All resources with corresponding sizes and operations are described in image-definitions.ts.

Added image-generation-test.png in order to provide easy way to test this functionality.

For image resizing we use https://github.com/oliver-moran/jimp.